Began dev of Master page which involves custom made .lua scripts
The Master tab/page is designed to support a Mastering-Desk-style-workflow.

This means that the mastering/mixer engineer will have the most important knobs and toggles in the chain in close reach in order to make quick decisions. This involves the Master track, the Stereo Bus or Mix bus, plus 4 extra buses for either Drums, Music, Vocals, Bass, etc.

These buses need to have their most important values exposed as low and high shelves, drive knobs, receive volumes, etc. Which would depend entirely on how the user sets them up.

The Master scripts help access fx parameters and on/off toggles as well as position of fxs in the chain in order to reduce the complexity of having all the parameters exposed (this means not all 24 parameters are exposed at the same time on the controller but you can set different plugins to reach for different fxparam within those). The Master tab switches to a 6 tracks bank with only 1 send, multiple receives, up to 12 fx per track and up to 24 fxparam per effect. Which helps the mastering engineer chose between fxs by turning them on and off as well as working on buses in parallel, etc.

The slider view enables this parallel workflow in order to avoid switching windows constantly.