Finished first part of Buses FX. Can reposition and toggle on/off fx using custom actions.
You can install the actions with the install_selected.lua scripts.

These actions are, ideally, temp, as Realearn could potentially implement these features.

For the FX in the Buses Page, the main idea is to have multi-gesture buttons for Open UI, Toggle On/Off, Hold for moving, etc.

This Page will become a two track, side to side style workflow in order to access fx parameters from two different tracks at the same time. The trick is that it won't have a 2 track bank, but rather, a 1 vs 2/3/4 style bank where you can select which one of the 4 tracks is show in the Left side and which one of the rest is shown in the Right side. This way you can work the master vs the mixbus or the master vs the drumbus, etc.

The idea is to be able to access small banks of parameters for each track and expose them in order to tweak them at the same time against other parameters from the other side.